+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased] - Date
+### Added
+
+### Fixed
+- Fix PDF preview not working due to a workaround
+
 ## [0.4.0] - 2022-06-20
 ### Added
 - NEW: Graphical User Interface: Download exe-file, - or run the file "gui.py" within the directory ".../transistordatabase/transistordatabase/gui" to start the GUI